requirements.txt
pdfplumber>=0.10
python-docx>=1.1
openai>=1.30

# Dev / test
pytest>=8.0
pytest-mock>=3.14
reportlab>=4.2